% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nifH_example_ps}
\alias{nifH_example_ps}
\title{Phyloseq object containing \emph{nifH} sequences analyzed in Kapili & Dekas, in prep.}
\format{\code{phyloseq-class} object with 1245 sequences and 13 samples.}
\source{
BJ Kapili and AE Dekas. PPIT: an R package for inferring microbial taxonomy from nifH sequences. In prep.
}
\usage{
nifH_example_ps
}
\description{
Phyloseq object containing 1245 \emph{nifH} sequences and 13 samples. For details on sequence prep, see Kapili & Dekas, in prep.
}
\keyword{datasets}
