% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nifH_example_SEPP_alignment}
\alias{nifH_example_SEPP_alignment}
\title{SEPP output alignment for PPIT tutorial.}
\format{\code{DNAMultipleAlignment} containing 10121 total sequences (8876 reference \emph{nifH} sequences; 1245 \emph{nifH} ASVs).}
\source{
BJ Kapili and AE Dekas. PPIT: an R package for inferring microbial taxonomy from nifH sequences. In prep.
}
\usage{
nifH_example_SEPP_alignment
}
\description{
SEPP output alignment to be used in PPIT tutorial. Produced by inserting the \emph{nifH} ASVs from \code{nifH_example_ps} into \code{nifH_reference_alignment_v2} using SEPP (v.4.3.5).
}
\keyword{datasets}
