% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nifH_reference_RAxML_info_v2}
\alias{nifH_reference_RAxML_info_v2}
\title{RAxML info file for SEPP placement (v1.0.0)}
\format{\code{Data frame} object with 11477 rows and 1 column.}
\source{
BJ Kapili and AE Dekas. PPIT: taxonomy inference for
nifH sequences accounting for known horizontal gene transfer events. In prep.
}
\usage{
nifH_reference_RAxML_info_v2
}
\description{
RAxML info file generated during construction of \code{nifH_reference_tree_v2}. To be used with SEPP (Mirarab et al., 2012) for insertion of \emph{nifH} query sequences
into reference \emph{nifH} alignment and tree.
}
\keyword{datasets}
