% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nifH_reference_tree_v2}
\alias{nifH_reference_tree_v2}
\title{Reference \emph{nifH} gene tree (v.1.2.0)}
\format{\code{phylo} object with 8876 tips.}
\source{
PPIT: an R package for inferring microbial taxonomy from nifH sequences
}
\usage{
nifH_reference_tree_v2
}
\description{
Reference \emph{nifH} gene tree to use for inserting \emph{nifH} query sequences. For details about tree construction,
see Kapili & Dekas, in prep.
}
\keyword{datasets}
