% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patristic.distance.R
\name{patristic.distance}
\alias{patristic.distance}
\title{Calculate patristic distances between query and reference sequences.}
\usage{
patristic.distance(query, tree)
}
\arguments{
\item{query}{Vector of query sequence names.}

\item{tree}{Phylogenetic tree in Newick format containing query sequences placed onto a reference tree.}
}
\value{
An object of class \code{"dist"}
}
\description{
Wrapper function that reformats the patristic distance output from the \code{cophenetic}
function in \code{ape}. Distance matrix output
reformatted to have query sequences as rows and reference sequences as columns.
}
