% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.identity.R
\name{percent.identity}
\alias{percent.identity}
\title{Calculate pairwise percent identity.}
\usage{
percent.identity(type, query, group, alignment_as_matrix)
}
\arguments{
\item{type}{Type of query sequence (i.e., "partial" if amplicon or "full" if full-length). For partial sequences,
pairwise percent identity calculated based on loci between first and last non-gap amplicon positions in the alignment.}

\item{query}{Name of the query sequence to compare to references.}

\item{group}{Vector of reference sequence names to compare to query sequence.}

\item{alignment}{Nucleotide alignment used for constructing phylogenetic tree of query and reference sequences (e.g., SEPP output alignment).}
}
\value{
An object of class \code{data frame}
}
\description{
Wrapper function that uses the \code{dist.gene} function in \code{ape} to calculate
pairwise percent identities between query and reference sequences. Loci in input alignment that only contain
gaps are removed prior to calculation.
}
