% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppit.R
\name{ppit}
\alias{ppit}
\title{Infer host identities of query sequences.}
\usage{
ppit(type, query, tree, alignment, taxonomy, opt_thresh, cutoffs)
}
\arguments{
\item{type}{Either "partial" or "full".}

\item{query}{Vector of query sequence names.}

\item{tree}{Phylogenetic tree in Newick format containing query sequences placed onto a reference tree.}

\item{alignment}{Nucleotide alignment used for constructing phylogenetic tree of query and reference sequences (e.g., SEPP output alignment).}

\item{opt_thresh}{Phylogenetic neighborhood threshold.}

\item{cutoffs}{Data frame containing patristic distance and pairwise percent identity cutoffs for each taxonomic rank.}
}
\value{
An object of class \code{data frame}
}
\description{
Function that infers taxonomic identity of marker gene source organisms based on phylogenetic placement and sequence identity. For a tutorial, type \code{Vignette('ppit')}.
}
