% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.partition.R
\name{tree.partition}
\alias{tree.partition}
\title{Determine optimal phylogenetic neighborhood threshold.}
\usage{
tree.partition(type, query, tree, alignment, taxonomy, cutoffs)
}
\arguments{
\item{query}{Vector of query sequence names.}

\item{tree}{Phylogenetic tree in Newick format containing query sequences placed onto a reference tree.}

\item{alignment}{Nucleotide alignment used for constructing phylogenetic tree of query and reference sequences (e.g., SEPP output alignment).}

\item{taxonomy}{Data frame containing taxonomic information and genomic location for each reference sequence.}

\item{cutoffs}{Data frame containing patristic distance and pairwise percent identity cutoffs for each taxonomic rank.}
}
\value{
An object of class \code{numeric}
}
\description{
Determines patristic distance threshold that maximizes the number of host inferences
returned at the phylum rank.
}
