% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nifH_reference_alignment_v2}
\alias{nifH_reference_alignment_v2}
\title{Reference \emph{nifH} nucleotide alignment (v.1.2.0)}
\format{\code{DNAMultipleAlignment} object with 8876 sequences}
\source{
BJ Kapili and AE Dekas. PPIT: an R package for inferring microbial taxonomy from nifH sequences. In prep.
}
\usage{
nifH_reference_alignment_v2
}
\description{
Reference \emph{nifH} nucleotide alignment to use for inserting \emph{nifH} query sequences. For details about alignment creation,
see Kapili & Dekas, in prep.
}
\keyword{datasets}
