% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nifH_reference_tree_v2}
\alias{nifH_reference_tree_v2}
\title{Reference \emph{nifH} gene tree (v1.0.0)}
\format{\code{phylo} object with 8539 tips.}
\source{
BJ Kapili and AE Dekas. PPIT: taxonomy inference for
nifH sequences accounting for known horizontal gene transfer events. In prep.
}
\usage{
nifH_reference_tree_v2
}
\description{
Reference \emph{nifH} gene tree to use for inserting \emph{nifH} query sequences. For details about tree construction,
see Kapili & Dekas, in prep.
}
\keyword{datasets}
